requirements.txt
Flask==1.1.2
python-dotenv
fuzzywuzzy==0.18.0
huggingface-hub==0.0.8
imbalanced-learn==0.8.0
jupyter==1.0.0
matplotlib==3.4.2
msgpack==1.0.0
nest-asyncio==1.5.1
networkx==2.6.2
nltk==3.6.2
numpy==1.19.5
pickleshare==0.7.5
prometheus-client==0.11.0
prompt-toolkit==3.0.19
pyenchant==3.2.0
python-Levenshtein-wheels==0.13.1
scikit-learn==0.24.2
scipy==1.4.1
simpletransformers==0.61.6
soupsieve==2.2.1
Sphinx==3.5.2
SQLAlchemy==1.3.22
subprocess32==3.5.4
sumy==0.8.1
tabulate==0.8.9
tokenizers==0.10.3
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tox==3.23.0
tqdm==4.49.0
